added the  new option to install required apps

+++ setup.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-while getopts "k:d:p:" opt; do
+while getopts "k:d:p:a:" opt; do
   case $opt in
     k)
       SSH_PRIVATE_KEY_PATH="$OPTARG"
@@ -10,6 +10,9 @@ while getopts "k:d:p:" opt; do
     p)
       PUB_KEY="$OPTARG"
       ;; 
+    a)
+      APP_NAME="$OPTARG"
+      ;;
     \?)
       echo "Invalid option: -$OPTARG" >&2
       exit 1
@@ -21,7 +24,7 @@ while getopts "k:d:p:" opt; do
   esac
 done
 
-if [[ -z $SSH_PRIVATE_KEY_PATH ]] || [[ -z $DISTRO ]] || [[ -z $PUB_KEY ]]; then
+if [[ -z $SSH_PRIVATE_KEY_PATH ]] || [[ -z $DISTRO ]] || [[ -z $PUB_KEY ]] || [[ -z $APP_NAME ]]; then
     echo "All -k (for key) and -d (for distro) and -pubkey (for public key) flags are required."
     exit 1
 fi
@@ -57,21 +60,40 @@ export EC2_IP=$(terraform output instance_public_ip)
 
 cat instance_info.txt
 ls
-cd ../jenkinsSetup
-pwd
 
 
-export EC2_IP
-export SSH_PRIVATE_KEY_PATH
-export EC2_USER
-rm inventory.yml
-echo "[linux]" > inventory.yml
-echo "$EC2_IP ansible_user=$EC2_USER ansible_ssh_private_key_file=$SSH_PRIVATE_KEY_PATH" >> inventory.yml
-echo "######Running ansible galaxy to install java package ######" 
-sleep 30
-ansible-galaxy install ansiblebit.oracle-java
-echo "######Running ansible playbook to setup jenkins ######"
-sleep 30
-ansible-playbook -i inventory.yml main.yml
-echo "###### Your Jenkins Initial Admin Password is ######"
-cat initialAdminPassword
+case $APP_NAME in 
+     "Jenkins")
+        echo "###### RUNNING JENKINS SETUP NOW #####"
+        cd ../jenkinsSetup
+        export EC2_IP
+        export SSH_PRIVATE_KEY_PATH
+        export EC2_USER
+        FILE="inventory.yml"
+        if [ -f "$FILE" ]; then
+            rm inventory.yml
+        else
+            echo "### Moving ### "
+        fi
+        
+        echo "[linux]" > inventory.yml
+        echo "$EC2_IP ansible_user=$EC2_USER ansible_ssh_private_key_file=$SSH_PRIVATE_KEY_PATH" >> inventory.yml
+        
+        sleep 30
+        
+        echo "######Running ansible playbook to setup jenkins ######"
+        sleep 30
+        ansible-playbook -i inventory.yml main.yml
+        echo "###### Your Jenkins Initial Admin Password is ######"
+        cat initialAdminPassword
+        rm initialAdminPassword
+        rm inventory.yml
+        ;;
+      "RabbitMQ")
+        echo "###### RUNNING RABBITMQ SETUP NOW #####"
+        ;;
+      *)
+        echo "Unsupported APP_NAME value"
+        exit 1
+        ;;
+esac
